banks/requirements.txt
json
requests
lxml.html
dnspython
tldextract
netaddr
scapy
unidecode
